make UNO enums scoped for internal LO code
this modifies codemaker so that, for an UNO enum, we generate code
that effectively looks like:

   #ifdef LIBO_INTERNAL_ONLY && HAVE_CX11_CONSTEXPR
       enum class XXX {
           ONE = 1
       };
       constexpr auto ONE = XXX_ONE;
   #else
      ...the old normal way..
   #endif

which means that for LO internal code, the enums are scoped.

The "constexpr auto" trick acts like an alias so we don't have to
use scoped naming everywhere.

+++ sd/qa/unit/export-tests-ooxml2.cxx
@@ -768,12 +768,12 @@ void SdOOXMLExportTest2::testTdf105739()
         // Test fill type
         drawing::FillStyle aFillStyle(drawing::FillStyle_NONE);
         aXBackgroundPropSet->getPropertyValue("FillStyle") >>= aFillStyle;
-        CPPUNIT_ASSERT_EQUAL(drawing::FillStyle_GRADIENT, aFillStyle);
+        CPPUNIT_ASSERT_EQUAL((int)drawing::FillStyle_GRADIENT, (int)aFillStyle);
 
         // Test gradient properties
         com::sun::star::awt::Gradient aFillGradient;
         aXBackgroundPropSet->getPropertyValue("FillGradient") >>= aFillGradient;
-        CPPUNIT_ASSERT_EQUAL(awt::GradientStyle_LINEAR, aFillGradient.Style);
+        CPPUNIT_ASSERT_EQUAL((int)awt::GradientStyle_LINEAR, (int)aFillGradient.Style);
         CPPUNIT_ASSERT_EQUAL(util::Color(0xff0000), aFillGradient.StartColor);
         CPPUNIT_ASSERT_EQUAL(util::Color(0x00b050), aFillGradient.EndColor);
     }
